Add support for streamline data
The new trk file format can support streamline data. This feature
is not supported by the streamlines package.

This commit adds partial support for streamline data associated
with points. For now, point data can be added to streamlines and
loaded/saved. Data associated with whole streamlines is not
supported yet.