ENH: refactor the streamlines script to load subcommands dynamically
The current streamlines script requires adding subparsers in the main
script for every subcommand. This makes it difficult to read and more
tedious to add new subcommands.

This commit start a refactor of the script by adding a way to load
subcommands dynamically. The merge subcommand is refactored using this
new approach. The other subcommands will be refactored in later
commits.