FEAT: preserve voxel size and shape on save and load
When an affine from voxel to native RAS is available in a .trk file, the
voxel size and shape of the reference image as also available. These
are currently not preserved by the save and load operation.

This commit modifies save and load to keep the voxel size and shape
information in the VoxelSpace coordinate system.